Restructured time series experiment files
