--- Time_Series/LTSF_models.py
@@ -1,6 +1,0 @@
-from tensorflow import keras
-
-
-def get_models(L=256):
-    return keras.Input(shape=(L,))
-
